backend/requirements.txt
django
djangorestframework
psycopg2-binary
cryptography
djangorestframework-simplejwt
drf-spectacular
django-filter
